Refresh access tokens with new auth endpoint
